Refactor import paths for base service in chatbot, text summarization, and code generation services
