+++ 02-code-generation-with-langchain/core/service/code_generation_service.py
@@ -23,8 +23,8 @@ import sys
 import os
 
 # Add the src directory to the path to import base_service
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../../../src")))
-from service.base_service import BaseGenerativeService
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../../../")))
+from src.service.base_service import BaseGenerativeService
 
 # Set up logger
 logger = logging.getLogger(__name__)
